+++ tests/test.nim
@@ -17,6 +17,7 @@ import ./t_unbounded_sipmuc
 import ./t_unbounded_sipmuc_threaded
 import ./t_unbounded_sipsic
 import ./t_unbounded_sipsic_threaded
+import ./t_unbounded_auto_create
 
 import ./t_wraparound
 
@@ -40,4 +41,5 @@ export
   t_unbounded_sipmuc_threaded,
   t_unbounded_sipsic,
   t_unbounded_sipsic_threaded,
+  t_unbounded_auto_create,
   t_wraparound